feat(IN-320): 홈페이지 SEO 적용
