--- app/(public)/page.ts
@@ -1,1 +1,0 @@
-export { HomePage as default } from '@/pages/home'
