+++ src/entities/home/featureCard/FeatureCard.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { Button } from '@/shared/ui/button'
